fix(app): register archive session command in both layouts
The session.archive command was registered inside LegacyLayout, which
never mounts when general.newLayoutDesigns is enabled (the default).
The keybind was dead and the command missing from the palette.

Move it to use-session-commands, mounted by the session route in both
shells, and extract the archive logic into a useSessionArchive hook
shared with the timeline menus and deleteSession.
